Ajustes no método de salvar as notas ficais no banco de dados e o tipo da coluna amount na base de dados

+++ src/main/resources/db/migration/V1.0__create_table_invoices.sql
@@ -3,7 +3,7 @@ USE template;
 CREATE TABLE invoices (
     id integer NOT NULL AUTO_INCREMENT,
     access_key varchar(200) NOT NULL,
-    amount decimal NOT NULL,
+    amount double(15,2) NOT NULL,
     PRIMARY KEY (id)
 );
